+++ src/components/FormPage/FillingForm/index.ts
@@ -1,0 +1,1 @@
+export { FillingForm } from './FillingForm.tsx';
